feat: extract EmbraceGoRouterObserver into standalone embrace_go_router package
