+++ embrace_go_router/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 1.0.0
+
+* Initial release of `embrace_go_router`. Provides `EmbraceGoRouterObserver` for automatic navigation tracking in apps using go_router.
