+++ embrace_go_router/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: package:very_good_analysis/analysis_options.3.0.2.yaml
