+++ embrace_go_router/lib/embrace_go_router.dart
@@ -1,0 +1,1 @@
+export 'src/go_router_observer.dart';
