+++ embrace_go_router/pubspec.yaml
@@ -1,0 +1,19 @@
+name: embrace_go_router
+description: Provides go_router navigation tracking for the Embrace SDK.
+version: 1.0.0
+homepage: https://embrace.io
+environment:
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
+dependencies:
+  embrace: ^4.6.0
+  flutter:
+    sdk: flutter
+  go_router: ">=6.0.0 <16.0.0"
+dev_dependencies:
+  embrace_platform_interface: ">=4.6.0 <4.7.0"
+  flutter_test:
+    sdk: flutter
+  mocktail: ">=0.3.0 <2.0.0"
+  plugin_platform_interface: ^2.0.0
+  very_good_analysis: ">=5.1.0 <11.0.0"
