+++ embrace_go_router/pubspec_overrides.yaml
@@ -1,0 +1,9 @@
+dependency_overrides:
+  embrace:
+    path: ../embrace
+  embrace_platform_interface:
+    path: ../embrace_platform_interface
+  embrace_android:
+    path: ../embrace_android
+  embrace_ios:
+    path: ../embrace_ios
